backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.4.0
